requirements.txt
# AlphaGPT - Python Dependencies
# Python 3.10+ required (recommended: 3.11)
#
# Installation:
#   pip3 install -r requirements.txt
#   # Or use: pip install -r requirements.txt (if pip points to Python 3)
#
# For experimental scripts (times.py, lord/experiment.py):
#   pip3 install -r requirements-optional.txt
#   # Or use: pip install -r requirements-optional.txt (if pip points to Python 3)

# ===== Core Dependencies (Required for main pipeline) =====

# Deep Learning & Numerical Computing
torch>=2.0.0          # PyTorch for model training and inference
numpy>=1.24.0         # Numerical operations

# Data Processing & Storage
pandas>=2.0.0         # Data manipulation and analysis
pyarrow>=14.0.0       # Parquet file read/write engine

# Crypto Exchange Data
ccxt>=4.0.0           # Unified crypto exchange API (Binance OHLCV + funding rates)

# Environment & Configuration
python-dotenv>=1.0.0  # Load .env files

# Logging & Progress
loguru>=0.7.0         # Structured logging
tqdm>=4.66.0          # Progress bars

# ===== Optional Dependencies (For experimental scripts) =====
# See requirements-optional.txt for:
#   - matplotlib, seaborn (for lord/experiment.py visualization)
#   - tushare (for times.py Chinese stock market data)